Upcase body before searching to find lowercase tracking numbers
The regexps define the search pattern in terms of uppercase characters,
so when the source data you're searching contains lowercase tracking
numbers, they don't match.

By upcasing the body before searching, it's more likely to match the
search patterns.

This also matches that initializing a new `TrackingNumber` object
upcases the number for its internal value.

Inspired by:
https://github.com/jkeen/tracking_number_data/pull/102

Meant as an interim solution until it's decided whether the source
regexps should be case-sensitive or not.